7Modules/Jelson/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
SQLAlchemy>=2.0.29
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
python-socketio>=5.11.2
pydantic[email]>=2.6.4
httpx>=0.27.0
